# Speed up JSON::Validator.validate