--- a/lib/json-schema/schema.rb
+++ b/lib/json-schema/schema.rb
@@ -11,7 +11,7 @@ module JSON
 
       # If there is an ID on this schema, use it to generate the URI
       if @schema['id'] && @schema['id'].kind_of?(String)
-        temp_uri = JSON::Util::URI.parse(@schema['id'])
+        temp_uri = Addressable::URI.parse(@schema['id'])
         if temp_uri.relative?
           temp_uri = uri.join(temp_uri)
         end