--- a/lib/json-schema/util/uri.rb
+++ b/lib/json-schema/util/uri.rb
@@ -4,7 +4,7 @@ module JSON
   module Util
     class URI
       include Singleton
-      attr_accessor :cache
+      attr_accessor :normalize_cache, :parse_cache
 
       def normalized_uri(uri)
         uri = Addressable::URI.parse(uri) unless uri.is_a?(Addressable::URI)
@@ -21,8 +21,13 @@ module JSON
       # This is OK because the worst case result of the race is
       # a cache miss
       def self.normalized_uri(uri)
-        instance.cache ||= {}
-        instance.cache[uri] ||= instance.normalized_uri(uri)
+        instance.normalize_cache ||= {}
+        instance.normalize_cache[uri] ||= instance.normalized_uri(uri)
+      end
+
+      def self.parse(uri)
+        instance.parse_cache ||= {}
+        instance.parse_cache[uri.to_s] ||= Addressable::URI.parse(uri)
       end
     end
   end