--- a/lib/json-schema/validator.rb
+++ b/lib/json-schema/validator.rb
@@ -138,15 +138,15 @@ module JSON
     end
 
     def absolutize_ref_uri(ref, parent_schema_uri)
-      ref_uri = Addressable::URI.parse(ref)
-      ref_uri.fragment = ''
+      ref_uri = JSON::Util::URI.parse(ref)
+      ref_uri.fragment = '' if !ref_uri.fragment.nil? && !ref_uri.fragment.empty?
 
       return ref_uri if ref_uri.absolute?
       # This is a self reference and thus the schema does not need to be re-loaded
       return parent_schema_uri if ref_uri.path.empty?
 
       uri = parent_schema_uri.clone
-      uri.fragment = ''
+      uri.fragment = '' if !uri.fragment.nil? && !uri.fragment.empty?
       Util::URI.normalized_uri(uri.join(ref_uri.path))
     end
 
@@ -344,7 +344,7 @@ module JSON
 
       def validator_for_uri(schema_uri)
         return default_validator unless schema_uri
-        u = Addressable::URI.parse(schema_uri)
+        u = JSON::Util::URI.parse(schema_uri)
         validator = validators["#{u.scheme}://#{u.host}#{u.path}"]
         if validator.nil?
           raise JSON::Schema::SchemaError.new("Schema not found: #{schema_uri}")
@@ -517,7 +517,7 @@ module JSON
       if schema.is_a?(String)
         begin
           # Build a fake URI for this
-          schema_uri = Addressable::URI.parse(fake_uuid(schema))
+          schema_uri = JSON::Util::URI.parse(fake_uuid(schema))
           schema = JSON::Schema.new(JSON::Validator.parse(schema), schema_uri, @options[:version])
           if @options[:list] && @options[:fragment].nil?
             schema = schema.to_array_schema
@@ -539,14 +539,14 @@ module JSON
             schema = self.class.schema_for_uri(schema_uri)
             if @options[:list] && @options[:fragment].nil?
               schema = schema.to_array_schema
-              schema.uri = Addressable::URI.parse(fake_uuid(schema.hash.to_s))
+              schema.uri = JSON::Util::URI.parse(fake_uuid(schema.hash.to_s))
               Validator.add_schema(schema)
             end
             schema
           end
         end
       elsif schema.is_a?(Hash)
-        schema_uri = Addressable::URI.parse(fake_uuid(schema.hash.to_s))
+        schema_uri = JSON::Util::URI.parse(fake_uuid(schema.hash.to_s))
         schema = JSON::Schema.stringify(schema)
         schema = JSON::Schema.new(schema, schema_uri, @options[:version])
         if @options[:list] && @options[:fragment].nil?