@mjc 2015-05-18
I'll update the pull request description shortly with new benchmark numbers, found a bug that was an even more dramatic speedup.