@mjc 2015-05-18
I also need to research finding ways to test concurrency and remove the mutex around schema initialization safely.