@mjc 2015-06-01
I'm pretty much done with this pass if we can't remove the mutex.