@RST-J 2015-06-02
To me that looks OK and I guess using some more memory is a good trade for increased performance.