@iainbeeston 2015-06-04
So it works when the key is a uri object but not when it is a string?