@mjc 2015-06-05
Do you think it'd be possible to use the stdlib's URI stuff?
http://ruby-doc.org/stdlib-2.2.2/libdoc/uri/rdoc/URI.html
On Thu, Jun 4, 2015 at 7:36 PM Michael Cohen <michael.joseph.cohen@gmail.com>
wrote:

> It looks like the objects get modified in incompatible ways sometimes? The
> error on Travis is proving hard to track down
>
> On Thu, Jun 4, 2015 at 3:19 PM Iain Beeston <notifications@github.com>
> wrote:
>
>> So it works when the key is a uri object but not when it is a string?
>>
>> —
>> Reply to this email directly or view it on GitHub
>> <https://github.com/ruby-json-schema/json-schema/pull/255#issuecomment-109055101>
>> .
>>
>