@iainbeeston 2015-12-23
@RST-J and @mjc Could you please take a look at #285 , which is my attempt to salvage this and make it releasable?